requirements.txt
openpyxl
reportlab
pillow
psycopg2-binary
tkcalendar
